Move common modules from muscle_manager into libmuscle
