+++ libmuscle/python/libmuscle/operator.py
